Goals

+++ internal/store/habits.go
@@ -14,12 +14,13 @@ const postCtxKey habitKey = "habit"
 type Habit struct {
 	ID         int64  `json:"id"`
 	Name       string `json:"name"`
-	UserID     int64  `json:"user_id"`
+	UserID     int64  `json:"-"`
 	Impact     string `json:"impact"`
+	GoalID     *int64 `json:"goal_id"`
 	Created_at string `json:"created_at"`
 	Updated_at string `json:"updated_at"`
 	Version    int    `json:"version"`
-	User       User   `json:"user"`
+	User       User   `json:"-"`
 }
 
 type HabitStore struct {
@@ -28,14 +29,14 @@ type HabitStore struct {
 
 func (s *HabitStore) Create(ctx context.Context, habit *Habit) error {
 	query := `
-    INSERT INTO habits (name, impact, user_id)
-    VALUES  ($1, $2, $3) RETURNING id, created_at, updated_at
+    INSERT INTO habits (name, impact, user_id, goal_id)
+    VALUES  ($1, $2, $3, $4) RETURNING id, created_at, updated_at
   `
 
 	ctx, cancel := context.WithTimeout(ctx, QueryTimeoutDuration)
 	defer cancel()
 
-	err := s.db.QueryRowContext(ctx, query, habit.Name, habit.Impact, habit.UserID).Scan(
+	err := s.db.QueryRowContext(ctx, query, habit.Name, habit.Impact, habit.UserID, habit.GoalID).Scan(
 		&habit.ID,
 		&habit.Created_at,
 		&habit.Updated_at,
@@ -48,11 +49,11 @@ func (s *HabitStore) Create(ctx context.Context, habit *Habit) error {
 	return nil
 }
 
-func (s *HabitStore) GetByID(ctx context.Context, id int64) (*Habit, error) {
+func (s *HabitStore) GetByID(ctx context.Context, id int64, userID int64) (*Habit, error) {
 	query := `
-    SELECT id, name, impact, created_at, updated_at, version
+		SELECT id, name, user_id, impact, goal_id, created_at, updated_at, version
     FROM habits
-    WHERE id = $1
+		WHERE id = $1 AND user_id = $2
   `
 
 	ctx, cancel := context.WithTimeout(ctx, QueryTimeoutDuration)
@@ -60,10 +61,12 @@ func (s *HabitStore) GetByID(ctx context.Context, id int64) (*Habit, error) {
 
 	var habit Habit
 
-	err := s.db.QueryRowContext(ctx, query, id).Scan(
+	err := s.db.QueryRowContext(ctx, query, id, userID).Scan(
 		&habit.ID,
 		&habit.Name,
+		&habit.UserID,
 		&habit.Impact,
+		&habit.GoalID,
 		&habit.Created_at,
 		&habit.Updated_at,
 		&habit.Version,
@@ -87,6 +90,7 @@ func (s *HabitStore) GetUserFeed(ctx context.Context, userID int64, fq Paginated
   		h.name,
       h.user_id,
      	h.impact,
+		  h.goal_id,
 		  h.created_at,
       h.version
 		FROM habits h
@@ -112,6 +116,7 @@ func (s *HabitStore) GetUserFeed(ctx context.Context, userID int64, fq Paginated
 			&h.Name,
 			&h.UserID,
 			&h.Impact,
+			&h.GoalID,
 			&h.Created_at,
 			&h.Version,
 		); err != nil {
@@ -124,13 +129,13 @@ func (s *HabitStore) GetUserFeed(ctx context.Context, userID int64, fq Paginated
 	return feed, nil
 }
 
-func (s *HabitStore) Delete(ctx context.Context, postID int64) error {
-	query := `DELETE FROM habits WHERE id = $1`
+func (s *HabitStore) Delete(ctx context.Context, postID int64, userID int64) error {
+	query := `DELETE FROM habits WHERE id = $1 AND user_id = $2`
 
 	ctx, cancel := context.WithTimeout(ctx, QueryTimeoutDuration)
 	defer cancel()
 
-	res, err := s.db.ExecContext(ctx, query, postID)
+	res, err := s.db.ExecContext(ctx, query, postID, userID)
 	if err != nil {
 		return err
 	}
@@ -147,18 +152,18 @@ func (s *HabitStore) Delete(ctx context.Context, postID int64) error {
 	return nil
 }
 
-func (s *HabitStore) Update(ctx context.Context, habit *Habit) error {
+func (s *HabitStore) Update(ctx context.Context, habit *Habit, userID int64) error {
 	query := `
 		UPDATE habits
-		SET name = $1, impact = $2, version = version + 1
-		WHERE id = $3 AND version = $4
+		SET name = $1, impact = $2, goal_id = $3, version = version + 1
+		WHERE id = $4 AND user_id = $5 AND version = $6
 		RETURNING version
 	`
 
 	ctx, cancel := context.WithTimeout(ctx, QueryTimeoutDuration)
 	defer cancel()
 
-	err := s.db.QueryRowContext(ctx, query, habit.Name, habit.Impact, habit.ID, habit.Version).Scan(&habit.Version)
+	err := s.db.QueryRowContext(ctx, query, habit.Name, habit.Impact, habit.GoalID, habit.ID, userID, habit.Version).Scan(&habit.Version)
 	if err != nil {
 		switch {
 		case errors.Is(err, sql.ErrNoRows):
